refactor(routing): #2152 Bundle C — [inviteToken] → [token] param migration
Dynamic-param vocabulary alignment per Phase B audit Section C: the
public guest-join URL slug is migrated from `[inviteToken]` to the
shorter, more general `[token]`. The "[token] = opaque secret" convention
already covers other public routes (e.g. accept-invite); this aligns the
join surface with that vocabulary.

## Scope (kept narrow on purpose)

- `apps/web/src/app/join/[inviteToken]/` → `apps/web/src/app/join/[token]/`
- `page.tsx` `params: { inviteToken }` → `params: { token }`
- `GuestJoinView` prop `inviteToken: string` → `token: string` (+ 2 internal
  usages — fetch URL build + useCallback deps array).
- `__tests__/guest-join.test.tsx` import path + 8 prop occurrences updated.

## Intentionally NOT touched (separate API-contract scope)

- `InviteTokenResponse.inviteToken` DTO field in `components/session/types.ts`
  — public API contract field. Backend response JSON keys are owned by the
  BE handler; renaming FE-side would require a coordinated `*Endpoints.cs`
  + DTO change, which is out of scope for this Bundle C URL migration.
- `SessionPlayerActionsEndpoints.cs` + `SessionQueryEndpoints.cs`
  `{inviteToken}` URL pattern params — same reason. The FE no longer
  references the BE param name; whatever the BE calls the slot is an
  implementation detail.
- `InviteSession.test.tsx` continues to construct `InviteTokenResponse`
  fixtures with the field name from the API DTO.

Towards #2152 (Bundle F + sub-issue spawn in follow-up commits).
