+++ apps/web/src/app/join/__tests__/guest-join.test.tsx
@@ -53,7 +53,7 @@ vi.mock('@/lib/api/core/secureStorage', () => ({
   clearEncryptionKey: () => undefined,
 }));
 
-import { GuestJoinView } from '../[inviteToken]/GuestJoinView';
+import { GuestJoinView } from '../[token]/GuestJoinView';
 
 // ─── Helpers ──────────────────────────────────────────────────────────────────
 
@@ -109,13 +109,13 @@ describe('GuestJoinView', () => {
 
   it('shows loading state initially', () => {
     setupFetch();
-    render(<GuestJoinView inviteToken="ABC123" />);
+    render(<GuestJoinView token="ABC123" />);
     expect(screen.getByTestId('loading-text')).toBeInTheDocument();
   });
 
   it('renders name input after session loads', async () => {
     setupFetch();
-    render(<GuestJoinView inviteToken="ABC123" />);
+    render(<GuestJoinView token="ABC123" />);
 
     await waitFor(() => {
       expect(screen.getByLabelText(/il tuo nome/i)).toBeInTheDocument();
@@ -127,7 +127,7 @@ describe('GuestJoinView', () => {
 
   it('disables submit button when name is empty', async () => {
     setupFetch();
-    render(<GuestJoinView inviteToken="ABC123" />);
+    render(<GuestJoinView token="ABC123" />);
 
     await waitFor(() => {
       expect(screen.getByLabelText(/il tuo nome/i)).toBeInTheDocument();
@@ -141,7 +141,7 @@ describe('GuestJoinView', () => {
   it('shows validation error for whitespace-only name via aria-label', async () => {
     setupFetch();
     const user = userEvent.setup();
-    render(<GuestJoinView inviteToken="ABC123" />);
+    render(<GuestJoinView token="ABC123" />);
 
     await waitFor(() => {
       expect(screen.getByLabelText(/il tuo nome/i)).toBeInTheDocument();
@@ -161,7 +161,7 @@ describe('GuestJoinView', () => {
   it('joins session and shows scoreboard after submitting valid name', async () => {
     setupFetch();
     const user = userEvent.setup();
-    render(<GuestJoinView inviteToken="ABC123" />);
+    render(<GuestJoinView token="ABC123" />);
 
     await waitFor(() => {
       expect(screen.getByLabelText(/il tuo nome/i)).toBeInTheDocument();
@@ -180,7 +180,7 @@ describe('GuestJoinView', () => {
 
   it('shows error state when session fetch fails', async () => {
     setupFetch({ sessionOk: false });
-    render(<GuestJoinView inviteToken="INVALID" />);
+    render(<GuestJoinView token="INVALID" />);
 
     await waitFor(() => {
       expect(screen.getByRole('heading', { name: /sessione non trovata/i })).toBeInTheDocument();
@@ -192,7 +192,7 @@ describe('GuestJoinView', () => {
     localStorage.setItem('improvvisata_guest_name', 'SavedUser');
     setupFetch({ tokenValid: true });
 
-    render(<GuestJoinView inviteToken="ABC123" />);
+    render(<GuestJoinView token="ABC123" />);
 
     await waitFor(() => {
       expect(screen.getByTestId('scoreboard')).toBeInTheDocument();
